backend/requirements.txt
fastapi==0.109.0
langchain_google_vertexai==0.1.0
# llama_index==0.9.41
git+https://github.com/jmugicagonz/llama_index.git
git+https://github.com/jmugicagonz/llama_index.git@grounding#subdirectory=llama-index-integrations/llms/llama-index-llms-vertex
llama-index-embeddings-langchain==0.1.2
protobuf==4.25.2
pycryptodome==3.20.0
pydantic==2.5.3
python_jose==3.3.0
uvicorn==0.27.0
cryptography==41.0.7
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.28
python-dotenv==1.0.1
chromadb==0.4.22
python-dotenv==1.0.1
google-cloud-discoveryengine==0.11.10
google-api-core==2.18.0
pyarrow==15.0.2